pyproject.toml
[project]
name = "blog"
version = "0.1.0"
description = "Мой первый проект 'Блог' на джанго"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.6",
    "python-decouple>=3.8",
]
